Add ship preview
Remove cell nodes in #cpuBoard and #playerBoard to generate them using
Javascript. Added events to ship type buttons to define what type of
ship the user wants, and added events to #playerBoard to highlight
cells to show a preview of how a new ship would look on the board.

Added a CSS rule to highlight buttons in _buttons.scss.