Add event listener to initialize battle
Added initializeGame function that will prepare the game interface to
start the battle. Added attackCpuBoard function which makes a the
player and cpu turn, declares who won the match, and creates a button
to reset the game.

Also modified the reset method of game.js to reset its private variable
_winnerMessage, and modified _board.scss to add styles to the container
of the reset button.