Add new ship symbols
Modified _ships and getBoardAndShips of gameboard.js to print different
symbols based on the type of ship, and modified receiveAttack to print
the same symbols but in uppercase.

Modified index.js to print the new symbols on the page and updated the
ships table to show what type of ship the each symbol represents.